+++ include/nf_text16.h
@@ -62,7 +62,8 @@ extern "C" {
 /// @param width Map width (in pixels)
 /// @param height Map height (in pixels)
 /// @param rotation Rotation (0 - 2)
-void NF_LoadTextFont16(const char *file, const char *name, u16 width, u16 height, u8 rotation);
+void NF_LoadTextFont16(const char *file, const char *name, u32 width, u32 height,
+                       u32 rotation);
 
 /// Create a special tiled background to write text on it with a 8x16 font.
 ///
@@ -80,7 +81,7 @@ void NF_LoadTextFont16(const char *file, const char *name, u16 width, u16 height
 /// @param layer Background layer (0 - 3)
 /// @param rotation Rotation (0 - 2)
 /// @param name Font name
-void NF_CreateTextLayer16(u8 screen, u8 layer, u8 rotation, const char *name);
+void NF_CreateTextLayer16(int screen, u32 layer, u32 rotation, const char *name);
 
 /// Write text in a layer with a 8x16 font at the specified coordinates.
 ///
@@ -111,7 +112,7 @@ void NF_CreateTextLayer16(u8 screen, u8 layer, u8 rotation, const char *name);
 /// @param x X coordinate
 /// @param y Y coordinate
 /// @param text String to write to the screen
-void NF_WriteText16(u8 screen, u8 layer, u16 x, u16 y, const char *text);
+void NF_WriteText16(int screen, u32 layer, u32 x, u32 y, const char *text);
 
 /// Clears the contents of a text layer with a 8x16 font filling it with zeroes.
 ///
@@ -123,7 +124,7 @@ void NF_WriteText16(u8 screen, u8 layer, u16 x, u16 y, const char *text);
 ///
 /// @param screen Screen (0 - 1)
 /// @param layer Background layer (0 - 3)
-void NF_ClearTextLayer16(u8 screen, u8 layer);
+void NF_ClearTextLayer16(int screen, u32 layer);
 
 /// @}
